requirements.txt
# Core
python-dotenv

# Jupyter
ipykernel
ipywidgets

# Data
pandas

# MongoDB
pymongo[srv]

# GCP
google-cloud-storage

